+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added Data validation. #21
 
+### Changed
+- Improved error message when a key is missing in `KeyedSheetDiff`. #23
+
 ### Fixed
 - Excluded the invalid odf dependency in pom.xml. This exclusion should be remove in the future. cdc-java/cdc-deps#17